% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dblp.R
\name{dblp_authors}
\alias{dblp_authors}
\title{Search DBLP authors}
\usage{
dblp_authors(query, hits = 30)
}
\arguments{
\item{query}{Author name search query}

\item{hits}{Maximum results (default 30)}
}
\value{
tibble: name, pid, url, notes
}
\description{
Search DBLP authors
}
