% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dblp.R
\name{dblp_venues}
\alias{dblp_venues}
\title{Search DBLP venues (conferences/journals)}
\usage{
dblp_venues(query, hits = 30)
}
\arguments{
\item{query}{Venue name search query}

\item{hits}{Maximum results (default 30)}
}
\value{
tibble: name, acronym, type, url
}
\description{
Search DBLP venues (conferences/journals)
}
