% Generated by build script
\name{dblp_venues}
\alias{dblp_venues}
\title{Search DBLP venues (conferences/journals)}
\usage{
dblp_venues()
}
\arguments{
\item{query}{Venue name search query}
\item{hits}{Maximum results (default 30)}
}
\value{tibble: name, acronym, type, url}
\description{Search DBLP venues (conferences/journals)}
